checking startup.sh file

+++ snmp_processor/mysql.go
@@ -15,12 +15,12 @@ func main() {
 	checkErr(err)
 
 	for rows.Next() {
-		var expid,runid,keyid,packets_sent,min_packet_length,max_packet_lenth int
-		var sampling_interval,min_intergramegap,max_intergramegap,status,when_to_process int
+		var expid,runid,keyid,packets_sent,min_packet_length,max_packet_lenth int32
+		var sampling_interval,min_intergramegap,max_intergramegap,status,when_to_process int32
 		var delay_on_shaper,packet_distribution,interframegap_distribution,destination string
 		err = rows.Scan(&expid,&runid,&keyid,&delay_on_shaper,&packets_sent,&min_packet_length,&max_packet_lenth,&packet_distribution,&sampling_interval,&min_intergramegap,&max_intergramegap,&interframegap_distribution,&destination,&status,&when_to_process)
 		checkErr(err)
-		unixtime := int64(time.Now().Unix())
+		unixtime := int32(time.Now().Unix())
 		if unixtime > when_to_process {
 			fmt.Println("ok, you can process")
 		}
